hello_openai/pyproject.toml
[project]
name = "hello-openai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httptools>=0.6.4",
    "openai>=1.52.2",
    "pip>=24.2",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.32.0",
    "uvloop>=0.21.0",
]
